Move test cases because they were annoying me
